src/model/back-requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
openai
faiss-cpu
python-dotenv
pydantic